feat & fix: new updates available
