+++ components/product/ProductGallery.tsx
@@ -1,4 +1,4 @@
-import ProductCard from "$store/islands/ProductCard.tsx";
+import ProductCard from "$store/components/product/ProductCard.tsx";
 import type {
   Layout as CardLayout,
 } from "$store/components/product/ProductCard.tsx";
