+++ components/product/ProductMatch.tsx
@@ -1,7 +1,7 @@
 import { SendEventOnLoad } from "$store/components/Analytics.tsx";
 import Image from "apps/website/components/Image.tsx";
 import Icon from "$store/components/ui/Icon.tsx";
-import ProductCard from "$store/islands/ProductCard.tsx";
+import ProductCard from "$store/components/product/ProductCard.tsx";
 import type {
   Layout as cardLayout,
 } from "$store/components/product/ProductCard.tsx";
