--- islands/ProductCard.tsx
@@ -1,8 +1,0 @@
-import Component from "$store/components/product/ProductCard.tsx";
-import type { Props } from "$store/components/product/ProductCard.tsx";
-
-function Island(props: Props) {
-  return <Component {...props} />;
-}
-
-export default Island;
